+++ next/.eslintrc.json
@@ -13,6 +13,7 @@
     "plugin:prettier/recommended"
   ],
   "rules": {
-    "react/react-in-jsx-scope": 0
+    "react/react-in-jsx-scope": 0,
+    "no-console": 0
   }
 }
