+++ next/pages/deck-builder.js
@@ -1,7 +1,11 @@
 import React, { useState, useCallback, useEffect } from 'react';
 import { useApolloClient } from 'react-apollo-hooks';
 import Layout from '../components/layout';
-import { initializeDeckBuilder } from '../lib/deck-utils';
+import {
+  initializeDeckBuilder as initDeck,
+  loadDeckFromSessionStorage,
+  loadDeckFromServer
+} from '../lib/deck-utils';
 import DeckBuilderSearchForm from '../components/deck-builder-search-form';
 import PageBanner from '../components/page-banner';
 import FactionFilters from '../components/faction-filters';
@@ -9,10 +13,10 @@ import DeckBuilderSidebar from '../components/deck-builder-sidebar';
 import CardSearchFilters from '../components/card-search-filters';
 import SliderSwitch from '../components/slider-switch';
 import DeckBuilderCardDisplay from '../components/deck-builder-card-display';
-import { singleDeckQuery, deckCardsQuery } from '../lib/deck-queries';
 import ErrorMessage from '../components/error-message';
+import PropTypes from 'prop-types';
 
-const initialSearchFilters = {
+const init = {
   cardSearchText: '',
   cardRarities: [],
   cardManaCosts: [],
@@ -20,94 +24,15 @@ const initialSearchFilters = {
   factions: []
 };
 
-const getInitialProps = async ({ query }) => {
-  // `null` instead of `undefined` to match sessionStorage.getItem('cantfind')
-  let deckId = parseInt(query.id, 10);
-  deckId = Number.isInteger(deckId) ? deckId : null;
-  return { deckId };
-};
-
-const loadDeckFromSessionStorage = setDeckInProgress => {
-  try {
-    const d = JSON.parse(sessionStorage.getItem('deckInProgress'));
-    if (!d) return;
-    const allProperties = Object.keys(initializeDeckBuilder());
-    const savedDeckProperties = Object.keys(d);
-    const isValidDeck = savedDeckProperties.reduce((isValid, key) => {
-      return isValid && allProperties.includes(key);
-    }, true);
-    if (isValidDeck) {
-      setDeckInProgress(d);
-      return true;
-    }
-  } catch (err) {
-    console.error(err);
-  }
-  return false;
-};
-
-const loadDeckFromServer = (
-  apolloClient,
-  deckId,
-  setDeckInProgress,
-  setIsError
-) => {
-  // noop if we don't have a legit deck id
-  if (!Number.isInteger(deckId)) return;
-
-  const deckPromise = apolloClient.query({
-    query: singleDeckQuery,
-    variables: {
-      id: deckId
-    }
-  });
-
-  const cardsPromise = apolloClient.query({
-    query: deckCardsQuery,
-    variables: {
-      id: deckId
-    }
-  });
-
-  Promise.all([deckPromise, cardsPromise])
-    .then(args => {
-      const deck = args[0].data.deck;
-      // I wonder if Lily can feel such a great disturbance in the force while
-      // she's on vacation :thinkingface:
-      const cards = args[1].data.deck.cardDecks.nodes;
-      setDeckInProgress({
-        deckName: deck.name,
-        deckPath: deck.path,
-        deckPower: deck.power,
-        mainDeck: cards.reduce((acc, c) => {
-          acc[c.card.id] = c;
-          return acc;
-        }, {})
-      });
-    })
-    .catch(reason => {
-      console.error(reason);
-      setIsError(true);
-    });
-};
-
 function DeckBuilderPage({ deckId }) {
-  const [cardSearchText, setCardSearchText] = useState(
-    initialSearchFilters.cardSearchText
-  );
-  const [cardRarities, setCardRarities] = useState(
-    initialSearchFilters.cardRarities
-  );
-  const [cardManaCosts, setCardManaCosts] = useState(
-    initialSearchFilters.cardManaCosts
-  );
-  const [supertypes, setSupertypes] = useState(initialSearchFilters.supertypes);
-  const [factions, setFactions] = useState(initialSearchFilters.factions);
+  const [cardSearchText, setCardSearchText] = useState(init.cardSearchText);
+  const [cardRarities, setCardRarities] = useState(init.cardRarities);
+  const [cardManaCosts, setCardManaCosts] = useState(init.cardManaCosts);
+  const [supertypes, setSupertypes] = useState(init.supertypes);
+  const [factions, setFactions] = useState(init.factions);
   const [currentTab, setTab] = useState('');
   const [viewFilters, setViewFilters] = useState(false);
-  const [deckInProgress, _setDeckInProgress] = useState(
-    initializeDeckBuilder()
-  );
+  const [deckInProgress, _setDeckInProgress] = useState(initDeck());
 
   const [isError, setIsError] = useState(false);
 
@@ -144,11 +69,11 @@ function DeckBuilderPage({ deckId }) {
   }, [deckId]);
 
   const handleClearFilters = useCallback(() => {
-    setCardSearchText(initialSearchFilters.cardSearchText);
-    setCardRarities(initialSearchFilters.cardRarities);
-    setCardManaCosts(initialSearchFilters.cardManaCosts);
-    setSupertypes(initialSearchFilters.supertypes);
-    setFactions(initialSearchFilters.factions);
+    setCardSearchText(init.cardSearchText);
+    setCardRarities(init.cardRarities);
+    setCardManaCosts(init.cardManaCosts);
+    setSupertypes(init.supertypes);
+    setFactions(init.factions);
   });
 
   return (
@@ -238,6 +163,15 @@ function DeckBuilderPage({ deckId }) {
   );
 }
 
-DeckBuilderPage.getInitialProps = getInitialProps;
+DeckBuilderPage.getInitialProps = async ({ query }) => {
+  // `null` instead of `undefined` to match sessionStorage.getItem('cantfind')
+  let deckId = parseInt(query.id, 10);
+  deckId = Number.isInteger(deckId) ? deckId : null;
+  return { deckId };
+};
+
+DeckBuilderPage.propTypes = {
+  deckId: PropTypes.number
+};
 
 export default DeckBuilderPage;
